model_interface: keep a pointer on evaluator in ModelCalcerHandler + properly synchronise it
This allow us to avoir a call to GetCurrentEvaluator at each Calc, that was causing issue due to the lock in it.
A pointer to the evaluator was previously present but got removed here, i kept the same approach but fixed it :
https://github.com/catboost/catboost/commit/db1801b50225cec8fe24ae9b314dc47050215ea6#diff-b27339156cee4c948ba57bc9c86da1ea719d70cf38a0eeef0b281fb44bed0718L17

My understanding is that those change are fully retro-compatible and shouldn't cause issues.
I only directly used the EVALUATOR_PTR in CalcFlat as it is the only thing i use, but it should be possible to use it in other places